refactor: localize 404 without initial props
