+++ public/locales/404.json
@@ -1,0 +1,14 @@
+{
+  "en": {
+    "goHomeButton": "Go Home",
+    "404error": "Error 404",
+    "hey": "Hey",
+    "404text": "We can’t seem to find the page you are looking for."
+  },
+  "ru": {
+    "goHomeButton": "На главную",
+    "404error": "Ошибка 404",
+    "hey": "Привет!",
+    "404text": "Мы не можем найти страницу, которую вы ищете."
+  }
+}
